Rename NewSlugInput
