+++ front/app/components/admin/SlugInput/index.tsx
